Funcion eliminar correctamente implementada

+++ src/app/teacher/menu-memory-game/menu-memory-game.ts
@@ -45,8 +45,19 @@ export class MenuMemoryGame implements OnInit, OnDestroy {
   }
 
   deleteGame() {
-    this.isDeleting = true;
-    this.selectedGameIndex = -1;
+    if (!this.isDeleting) {
+      // Primera vez: activar modo selección
+      this.isDeleting = true;
+      this.selectedGameIndex = -1;
+    } else {
+      // Segunda vez: eliminar si hay algo seleccionado
+      if (this.selectedGameIndex !== -1 && this.selectedGameIndex < this.games.length) {
+        console.log('Eliminando juego en índice:', this.selectedGameIndex);
+        this.gameService.deleteGame(this.selectedGameIndex);
+        this.isDeleting = false;
+        this.selectedGameIndex = -1;
+      }
+    }
   }
 
   selectGame(index: number) {
@@ -61,7 +72,8 @@ export class MenuMemoryGame implements OnInit, OnDestroy {
   }
 
   confirmDelete() {
-    if (this.selectedGameIndex !== -1) {
+    if (this.selectedGameIndex !== -1 && this.selectedGameIndex < this.games.length) {
+      console.log('Eliminando juego en índice:', this.selectedGameIndex);
       this.gameService.deleteGame(this.selectedGameIndex);
       this.isDeleting = false;
       this.selectedGameIndex = -1;
